+++ src/index.ts
@@ -92,6 +92,7 @@ import { omitIndexHtml } from './rules/omit-index-html';
 import { enforceIdCapitalization } from './rules/enforce-id-capitalization';
 import { noUnusedUseState } from './rules/no-unused-usestate';
 import { noUuidv4Base62AsKey } from './rules/no-uuidv4-base62-as-key';
+import enforceDynamicFileNaming from './rules/enforce-dynamic-file-naming';
 
 module.exports = {
   meta: {
@@ -206,6 +207,7 @@ module.exports = {
         '@blumintinc/blumint/omit-index-html': 'error',
         '@blumintinc/blumint/enforce-id-capitalization': 'error',
         '@blumintinc/blumint/no-unused-usestate': 'error',
+        '@blumintinc/blumint/enforce-dynamic-file-naming': 'error',
       },
     },
   },
@@ -236,6 +238,7 @@ module.exports = {
     'no-unused-props': noUnusedProps,
     'no-useless-fragment': noUselessFragment,
     'no-uuidv4-base62-as-key': noUuidv4Base62AsKey,
+    'enforce-dynamic-file-naming': enforceDynamicFileNaming,
     'prefer-fragment-shorthand': preferFragmentShorthand,
     'prefer-type-over-interface': preferTypeOverInterface,
     'require-memo': requireMemo,